tests/wpt/web-platform-tests/workers/Worker_ErrorEvent_filename.htm

<!DOCTYPE html>
<title> AbstractWorker ErrorEvent.filename </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/ErrorEvent.js');
  worker.onerror = t.step_func_done(function(e) {
    var href = location.href;
    var expected = href.substring(0, href.lastIndexOf('/')) + '/support/ErrorEvent.js';
    assert_equals(e.filename, expected);
  });
  worker.postMessage("Error Message");
});
</script>
